rename plot and tests directories
