+++ python/src/scipp/__init__.py
@@ -16,7 +16,7 @@ from . import detail
 from . import neutron
 from .show import show, make_svg
 from .table import table
-from . import plot
+from .plotting import plot
 from .extend_units import *
 from .table_html import to_html, make_html
 from .object_list import _repr_html_
@@ -75,9 +75,9 @@ from ._bins import _groupby_bins
 setattr(GroupByDataArray, 'bins', property(_groupby_bins))
 setattr(GroupByDataset, 'bins', property(_groupby_bins))
 
-setattr(Variable, 'plot', plot.plot)
-setattr(VariableConstView, 'plot', plot.plot)
-setattr(DataArray, 'plot', plot.plot)
-setattr(DataArrayConstView, 'plot', plot.plot)
-setattr(Dataset, 'plot', plot.plot)
-setattr(DatasetConstView, 'plot', plot.plot)
+setattr(Variable, 'plot', plot)
+setattr(VariableConstView, 'plot', plot)
+setattr(DataArray, 'plot', plot)
+setattr(DataArrayConstView, 'plot', plot)
+setattr(Dataset, 'plot', plot)
+setattr(DatasetConstView, 'plot', plot)
