+++ python/src/scipp/plotting/controller.py
