+++ python/src/scipp/plotting/controller1d.py
