+++ python/src/scipp/plotting/controller2d.py
