+++ python/src/scipp/plotting/controller3d.py
