+++ python/src/scipp/plotting/dispatch.py
