+++ python/src/scipp/plotting/figure.py
