+++ python/src/scipp/plotting/figure1d.py
