+++ python/src/scipp/plotting/figure2d.py
