+++ python/src/scipp/plotting/figure3d.py
