+++ python/src/scipp/plotting/helpers.py
