+++ python/src/scipp/plotting/model.py
