+++ python/src/scipp/plotting/model1d.py
