+++ python/src/scipp/plotting/model2d.py
