+++ python/src/scipp/plotting/model3d.py
