+++ python/src/scipp/plotting/panel.py
