+++ python/src/scipp/plotting/panel1d.py
