+++ python/src/scipp/plotting/panel3d.py
