+++ python/src/scipp/plotting/plot.py
