+++ python/src/scipp/plotting/plot1d.py
