+++ python/src/scipp/plotting/plot2d.py
