+++ python/src/scipp/plotting/plot3d.py
