+++ python/src/scipp/plotting/profile.py
