+++ python/src/scipp/plotting/resampling_model.py
