+++ python/src/scipp/plotting/sciplot.py
