+++ python/src/scipp/plotting/toolbar.py
