+++ python/src/scipp/plotting/tools.py
