+++ python/src/scipp/plotting/view.py
