+++ python/src/scipp/plotting/view1d.py
