+++ python/src/scipp/plotting/view2d.py
