+++ python/src/scipp/plotting/view3d.py
