+++ python/src/scipp/plotting/widgets.py
