+++ python/tests/plotting/test_plot_methods.py
