+++ python/tests/plotting/test_plot_nd.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 import scipp as sc
 from plot_helper import make_dense_dataset
-from scipp.plot import plot
+from scipp import plot
 
 # TODO: For now we are just checking that the plot does not throw any errors.
 # In the future it would be nice to check the output by either comparing
@@ -44,13 +44,6 @@ def test_plot_sliceviewer_2d_with_attrs():
 
 def test_plot_sliceviewer_2d_with_binedges():
     plot(make_dense_dataset(ndim=3, binedges=True))
-
-
-def test_plot_convenience_methods():
-    d = make_dense_dataset(ndim=3)
-    sc.plot.superplot(d)
-    sc.plot.image(d)
-    sc.plot.scatter3d(d)
 
 
 def test_plot_variable_3d():
